user/vitaly_m/KazImageScraper/requirements.txt
selenium==3.141.0
requests==2.25.1
pillow==9.0.1
pandas
python-dotenv